+++ srb/src/main/java/com/roi/controller/admin/EditStudentController.java
@@ -1,5 +1,6 @@
 package com.roi.controller.admin;
 
+import com.roi.controller.MainController;
 import com.roi.entity.Student;
 import com.roi.entity.Year;
 import com.roi.repository.*;
@@ -19,6 +20,8 @@ import java.util.Map;
 @Controller
 @RequestMapping("/admin/studentsList")
 public class EditStudentController {
+    @Autowired
+    private MainController mainController;
 
     @Autowired
     private StudentRepository studentRepository;
@@ -82,14 +85,15 @@ public class EditStudentController {
 
     @RequestMapping(value = {"/edit/{id}"}, method = RequestMethod.GET)
     public ModelAndView editStudentPage(@PathVariable Integer id) {
-
-        ModelAndView model = new ModelAndView();
-        Student student=studentRepository.findOne(id);
-        model.addObject("student",student);
-        model.addObject("error",ERROR_MESSAGE);
-        ERROR_MESSAGE=null;
-        model.setViewName("admin-page/student-service/edit-student");
-        return model;
+        if(studentRepository.exists(id)) {
+            ModelAndView model = new ModelAndView();
+            Student student = studentRepository.findOne(id);
+            model.addObject("student", student);
+            model.addObject("error", ERROR_MESSAGE);
+            ERROR_MESSAGE = null;
+            model.setViewName("admin-page/student-service/edit-student");
+            return model;
+        }else return mainController.errorPage();
     }
 
     @RequestMapping(value = {"/edit/{id}"}, method = RequestMethod.POST)
@@ -98,20 +102,22 @@ public class EditStudentController {
                               @RequestParam("login") String loginStr,
                               @RequestParam("password") String password,
                               @RequestParam("year") String yearName ){
-        Integer login=Integer.parseInt(loginStr);
-        Year year=yearRepository.findByName(Integer.parseInt(yearName));
-        Student student=studentRepository.findOne(id);
-        boolean  enable=student.getLogin().equals(login)||studentRepository.findByLogin(login)==null;
-        if(enable) {
-           student.setName(name);
-           student.setYear(year);
-           student.setPassword(password);
-           studentRepository.save(student);
-           return "redirect:/admin/studentsList";
-       } else {
-           ERROR_MESSAGE = "Такой логин уже существует!";
-           return "redirect:/admin/studentsList/edit/{id}";
-       }
+        if(studentRepository.exists(id)) {
+            Integer login = Integer.parseInt(loginStr);
+            Year year = yearRepository.findByName(Integer.parseInt(yearName));
+            Student student = studentRepository.findOne(id);
+            boolean enable = student.getLogin().equals(login) || studentRepository.findByLogin(login) == null;
+            if (enable) {
+                student.setName(name);
+                student.setYear(year);
+                student.setPassword(password);
+                studentRepository.save(student);
+                return "redirect:/admin/studentsList";
+            } else {
+                ERROR_MESSAGE = "Такой логин уже существует!";
+                return "redirect:/admin/studentsList/edit/{id}";
+            }
+        }else return "error";
     }
 
 
@@ -119,20 +125,24 @@ public class EditStudentController {
 
     @RequestMapping(value = {"/delete/{id}"}, method = RequestMethod.GET)
     public ModelAndView deleteStudentPage(@PathVariable Integer id) {
-        ModelAndView model = new ModelAndView();
-        Student student=studentRepository.findOne(id);
-        model.addObject("student",student);
-        model.setViewName("admin-page/student-service/delete-student");
-        return model;
+        if(studentRepository.exists(id)) {
+            ModelAndView model = new ModelAndView();
+            Student student = studentRepository.findOne(id);
+            model.addObject("student", student);
+            model.setViewName("admin-page/student-service/delete-student");
+            return model;
+        } else return mainController.errorPage();
     }
 
     @RequestMapping(value = {"/delete/{id}"}, method = RequestMethod.POST)
     public String deleteStudent(@PathVariable Integer id){
-        Student student=studentRepository.findOne(id);
-        sessionUtils.expireUserSessions("st"+student.getLogin());
-        markRepository.removeByStudent(student);
-        studentRepository.removeById(id);
-        return "redirect:/admin/studentsList";
+        if (studentRepository.exists(id)) {
+            Student student = studentRepository.findOne(id);
+            sessionUtils.expireUserSessions("st" + student.getLogin());
+            markRepository.removeByStudent(student);
+            studentRepository.removeById(id);
+            return "redirect:/admin/studentsList";
+        }else return "error";
     }
 
 }
