+++ srb/src/main/java/com/roi/controller/admin/EditTeacherController.java
@@ -1,6 +1,7 @@
 package com.roi.controller.admin;
 
 
+import com.roi.controller.MainController;
 import com.roi.entity.Subject;
 import com.roi.entity.Teacher;
 import com.roi.repository.*;
@@ -21,6 +22,8 @@ import java.util.Map;
 @RequestMapping("/admin/teachersList")
 public class EditTeacherController {
 
+    @Autowired
+    private MainController mainController;
 
     @Autowired
     private SessionUtils sessionUtils;
@@ -81,13 +84,15 @@ public class EditTeacherController {
 
     @RequestMapping(value = {"/edit/{id}"}, method = RequestMethod.GET)
     public ModelAndView editTeacherPage(@PathVariable Integer id) {
-        ModelAndView model = new ModelAndView();
-        Teacher teacher=teacherRepository.findOne(id);
-        model.addObject("teacher",teacher);
-        model.addObject("error",ERROR_MESSAGE);
-        ERROR_MESSAGE=null;
-        model.setViewName("admin-page/teacher-service/edit-teacher");
-        return model;
+        if(teacherRepository.exists(id)) {
+            ModelAndView model = new ModelAndView();
+            Teacher teacher = teacherRepository.findOne(id);
+            model.addObject("teacher", teacher);
+            model.addObject("error", ERROR_MESSAGE);
+            ERROR_MESSAGE = null;
+            model.setViewName("admin-page/teacher-service/edit-teacher");
+            return model;
+        }else return mainController.errorPage();
     }
 
     @RequestMapping(value = {"/edit/{id}"}, method = RequestMethod.POST)
@@ -95,45 +100,51 @@ public class EditTeacherController {
                               @RequestParam("teacherName") String name,
                               @RequestParam("login") String loginStr,
                               @RequestParam("password") String password) {
-        Integer login=Integer.parseInt(loginStr);
-        Teacher teacher=teacherRepository.findOne(id);
-        boolean  enable=teacher.getLogin().equals(login)||teacherRepository.findByLogin(login)==null;
+        if(teacherRepository.exists(id)) {
+            Integer login = Integer.parseInt(loginStr);
+            Teacher teacher = teacherRepository.findOne(id);
+            boolean enable = teacher.getLogin().equals(login) || teacherRepository.findByLogin(login) == null;
 
-        if(enable) {
-        teacher.setLogin(login);
-        teacher.setName(name);
-        teacher.setPassword(password);
-        teacherRepository.save(teacher);
-        return "redirect:/admin/teachersList";
-        } else {
-            ERROR_MESSAGE = "Такой логин уже существует!";
-            return "redirect:/admin/teachersList/edit/{id}";
-        }
+            if (enable) {
+                teacher.setLogin(login);
+                teacher.setName(name);
+                teacher.setPassword(password);
+                teacherRepository.save(teacher);
+                return "redirect:/admin/teachersList";
+            } else {
+                ERROR_MESSAGE = "Такой логин уже существует!";
+                return "redirect:/admin/teachersList/edit/{id}";
+            }
+        }else return "error";
     }
 
     //Удаление преподавателя
 
     @RequestMapping(value = {"/delete/{id}"}, method = RequestMethod.GET)
     public ModelAndView deleteTeacherPage(@PathVariable Integer id) {
-        ModelAndView model = new ModelAndView();
-        Teacher teacher=teacherRepository.findOne(id);
-        model.addObject("teacher",teacher);
-        model.setViewName("admin-page/teacher-service/delete-teacher");
-        return model;
+        if(teacherRepository.exists(id)) {
+            ModelAndView model = new ModelAndView();
+            Teacher teacher = teacherRepository.findOne(id);
+            model.addObject("teacher", teacher);
+            model.setViewName("admin-page/teacher-service/delete-teacher");
+            return model;
+        }else return mainController.errorPage();
     }
 
     @RequestMapping(value = {"/delete/{id}"}, method = RequestMethod.POST)
-    public String deleteTeacher(@PathVariable Integer id){
-        Teacher teacher=teacherRepository.findById(id);
-        sessionUtils.expireUserSessions("te"+teacher.getLogin());
-        List<Subject> subjectList=subjectRepository.findByTeacher(teacher);
+    public String deleteTeacher(@PathVariable Integer id) {
+        if (teacherRepository.exists(id)) {
+            Teacher teacher = teacherRepository.findById(id);
+            sessionUtils.expireUserSessions("te" + teacher.getLogin());
+            List<Subject> subjectList = subjectRepository.findByTeacher(teacher);
 
-        for(Subject s: subjectList){
-            s.setTeacher(null);
-            subjectRepository.save(s);
-        }
+            for (Subject s : subjectList) {
+                s.setTeacher(null);
+                subjectRepository.save(s);
+            }
 
-        teacherRepository.removeById(id);
-        return "redirect:/admin/teachersList";
+            teacherRepository.removeById(id);
+            return "redirect:/admin/teachersList";
+        }else return "error";
     }
 }
